+++ packages/runtime/src/interpreter/interpreter.ts
@@ -37,6 +37,10 @@ export class Interpreter {
   instructions: Operator[];
   instructionIndex: number;
   runtime: Runtime;
+  // The call stack is separate from the data stack. Only callsub and retsub manipulate it.
+  // It is used to provide sub routine functionality
+  callStack: Stack<number>;
+  labelMap: Map<string, number>; // label string mapped to their respective indexes in instructions array
 
   constructor () {
     this.stack = new Stack<StackElem>();
@@ -53,6 +57,8 @@ export class Interpreter {
     this.instructions = [];
     this.instructionIndex = 0; // set instruction index to zero
     this.runtime = <Runtime>{};
+    this.callStack = new Stack<number>();
+    this.labelMap = new Map<string, number>();
   }
 
   /**
@@ -265,6 +271,7 @@ export class Interpreter {
   /**
    * Description: moves instruction index to "label", throws error if label not found
    * @param label: branch label
+   * @param line: line number
    */
   jumpForward (label: string, line: number): void {
     while (++this.instructionIndex < this.instructions.length) {
@@ -285,6 +292,30 @@ export class Interpreter {
   }
 
   /**
+   * Description: moves instruction index to "label", throws error if label not found
+   * @param label: branch label
+   * @param line: line number
+   */
+  jumpToLabel (label: string, line: number): void {
+    const toInstructionIndex = this.labelMap.get(label);
+    if (toInstructionIndex === undefined) {
+      throw new RuntimeError(RUNTIME_ERRORS.TEAL.LABEL_NOT_FOUND, {
+        label: label,
+        line: line
+      });
+    }
+    let currentIndex = toInstructionIndex;
+    // if next immediate op is also label, then keep continuing, otherwise return
+    for (; currentIndex < this.instructions.length - 1; ++currentIndex) {
+      const nextInstruction = this.instructions[currentIndex + 1];
+      if (!(nextInstruction instanceof Label)) {
+        this.instructionIndex = currentIndex;
+        break;
+      }
+    }
+  }
+
+  /**
    * logs TEALStack upto depth = debugStack to console
    * @param instruction interpreter opcode instance
    * @param debugStack max no. of elements to print from top of stack
@@ -302,7 +333,17 @@ export class Interpreter {
   }
 
   /**
-   * Assets transaction references (apps, assets, accounts) lengths are valid:
+   * Maps labels with indexes according to instructions array
+   */
+  mapLabelWithIndexes (): void {
+    this.instructions.forEach((instruction, idx) => {
+      if (instruction instanceof Label) {
+        this.labelMap.set(instruction.label, idx);
+      }
+    });
+  }
+
+  /* Assets transaction references (apps, assets, accounts) lengths are valid:
    * 1. Application args are limited to max. size of 16.
    * 2. The AVM limits the accounts array to no more than 4
    * 3. Assets and application arrays combined and totaled with the accounts array can not exceed 8
@@ -348,6 +389,7 @@ export class Interpreter {
   execute (program: string, mode: ExecutionMode, runtime: Runtime, debugStack?: number): void {
     this.runtime = runtime;
     this.instructions = parser(program, mode, this);
+    this.mapLabelWithIndexes();
     if (mode === ExecutionMode.APPLICATION) { this.assertValidTxArray(); }
 
     let dynamicCost = 0;